refactor: Improve input and select UI components
Make the `label` prop optional in `Input` and `Select` components. This allows for inputs without visible labels, improving flexibility.

Add conditional rendering for labels in both components and adjust the styling for inputs that do not have a label to prevent unnecessary top margin.

Also, fixed a missing border style on the `Select` component.